pyproject.toml
[project]
name = "klarsicht"
version = "0.1.0"
description = "AI-powered RCA agent for Kubernetes"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "kubernetes>=31.0.0",
    "requests>=2.32.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.7.0",
    "langchain>=0.3.0",
    "langchain-anthropic>=0.3.0",
    "langchain-openai>=0.3.0",
    "langchain-core>=0.3.0",
    "asyncpg>=0.30.0",
    "python-jose[cryptography]>=3.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.25.0",
    "httpx>=0.28.0",
]
watsonx = [
    "langchain-ibm>=0.3.0",
]

[tool.setuptools.packages.find]
include = ["app*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
